???
How did this get here? This was supposed to be part of the last commit.

Oh, and the build is stable again, it just doesn't have all the features
it last had when it was in master.  Safe to compile, should run,
shouldn't segfault. This will stay true again until otherwise stated.