Message handler for Sprite interaction, RNG, sprite AI
Added a message handler for Sprite interaction. Used message handler to
finish implementing the ability to talk to a Sprite to receive a full
heal.

RNG is now a function of Party, and has been implemented using C++11
functions. Global RNG removed.

Tested basic AI for Sprites ('Walk SOUTH on one-second timer
intervals'). Implementation incomplete  -- the gameplay becomes very
jerky when this is turned on. For the purpose of this upload, reset
Sprites to stand still again.