Timer, NPC movement tests
Timer now ticks every 1/8th second.  Movement is queued on input, but is
now only executed on the timer.

Changed 'KillTest' NPC to seek a warp tile and then path to it, removing
itself from the map in that manner instead of just disappearing.

NPCs as described in map files now also must have minimum and maximum
random movement frequency. A minimum move frequency of 0 indicates a
stationary NPC.