Leveling, Larger Party Support
When leveling at the end of a battle, the text now informs you of such.
As well, current HP increases at the same time as maximum HP (so you'd
level from, say, 950/1000 HP to 1050/1100). This also holds true for MP,
though it isn't yet used or displayed.

Battle screen now displays information for all members of both the enemy
and player parties.