Turning now a complete MOVE
When told to move in a direction, the initial move is to just change
direction instead of moving to a different tile now.

NPCs receive a second move command in rapid succession so they move in
the same manner. The party requires a delay period between turning and
beginning to move in the new direction in order to implement
non-auto-turning, which has not yet been added.